# refactor: prototype reducing iterator overhead in swaps