--- a/x/concentrated-liquidity/swapstrategy/one_for_zero.go
+++ b/x/concentrated-liquidity/swapstrategy/one_for_zero.go
@@ -148,7 +148,16 @@ func (s oneForZeroStrategy) ComputeSwapStepInGivenOut(sqrtPriceCurrent, sqrtPric
 	return sqrtPriceNext, amountZeroOut, amountOneIn, feeChargeTotal
 }
 
-// TODO: spec
+// InitializeTickIterator returns iterator that seeks to the given tickIndex.
+// If tick index does not exist in the store, it will return an invalid iterator.
+//
+// oneForZeroStrategy assumes moving to the right of the current square root price.
+// As a result, we use forward iterator in InitializeTickIterator to find the next
+// tick to the left of current. The end cursor for forward iteration is inclusive.
+// Therefore, it searches directly from the tickIndex forwards in increasing lexicographic order.
+// Returns an invalid iterator if tickIndex is not in the store.
+// Panics if fails to parse tick index from bytes.
+// The caller is responsible for closing the iterator on success.
 func (s oneForZeroStrategy) InitializeTickIterator(ctx sdk.Context, poolId uint64, tickIndex int64) dbm.Iterator {
 	store := ctx.KVStore(s.storeKey)
 	prefixBz := types.KeyTickPrefixByPoolId(poolId)
@@ -176,7 +185,7 @@ func (s oneForZeroStrategy) InitializeTickIterator(ctx sdk.Context, poolId uint6
 // on the actual current tick.
 //
 // oneForZeroStrategy assumes moving to the right of the current square root price.
-// As a result, we use forward iterator in NextInitializedTick to find the next
+// As a result, we use forward iterator in InitializeTickIterator to find the next
 // tick to the left of current. The end cursor for forward iteration is inclusive.
 // Therefore, this method is, essentially a no-op. The logic is reversed for
 // zeroForOneStrategy where we use reverse iterator and have to add one to