… x/concentrated-liquidity/swapstrategy/zero_for_one_test.go