<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
v    If your PR doesn't close an issue, that's OK!  Just remove the Closes: #XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

Closes: #XXX

## What is the purpose of the change

This change reduces the overhead stemming from re-initrializing the iterator while performing a swap.
Instead, this change reuses the sam iterator across swap steps.

It is currently in the prototype state and needs to be cleaned up for the PR to be ready. However, the result shows a significant improvement relative to what we had prior to starting benchmarking:
```
benchstat old.bench new2.bench
old.bench:5: missing iteration count
new2.bench:5: missing iteration count
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/osmosis/v15/x/concentrated-liquidity
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                     │  old.bench   │               new2.bench               │
                     │    sec/op    │    sec/op     vs base                  │
SwapExactAmountIn-16   777.9m ± 23%   343.6m ± 26%  -55.84% (p=0.000 n=10+6)
```

The benchmark above included the following improvements as well:
- https://github.com/osmosis-labs/osmosis/pull/5176
- https://github.com/osmosis-labs/osmosis/pull/5171
- https://github.com/osmosis-labs/osmosis/pull/5174


Notice that the overhead from iterators in pprof flame graphs is completely eliminated:

Old:
![image](https://github.com/osmosis-labs/osmosis/assets/34196718/5408702d-0e35-47b5-a374-5c29820e1cfc)


New:
![image](https://github.com/osmosis-labs/osmosis/assets/34196718/15b8cded-677a-4d58-858f-aec37c538bbe)


## Testing and Verifying

*(Please pick one of the following options)*

This change is a trivial rework / code cleanup without any test coverage.

*(or)*

This change is already covered by existing tests, such as *(please describe tests)*.

*(or)*

This change added tests and can be verified as follows:

*(example:)*
  - *Added unit test that validates ...*
  - *Added integration tests for end-to-end deployment with ...*
  - *Extended integration test for ...*
  - *Manually verified the change by ...*

## Documentation and Release Note

  - [ ] Does this pull request introduce a new feature or user-facing behavior changes?
  - [ ] Changelog entry added to `Unreleased` section of `CHANGELOG.md`?

Where is the change documented? 
  - [ ] Specification (`x/{module}/README.md`)
  - [ ] Osmosis documentation site
  - [ ] Code comments?
  - [ ] N/A