@ValarDragon 2023-05-13
Awesome! One minor note, the iterator overhead isn't entirely eliminated though. Its in the `Next` calls, and in another process

The CacheKV overhead for iterator instantiation is completely eliminated though!